ResizingTextEdit sourcefile
