+++ src/widgets/resizingtextedit.hpp
@@ -2,50 +2,24 @@
 
 #include <QKeyEvent>
 #include <QTextEdit>
+#include <QCompleter>
 #include <boost/signals2.hpp>
 
 class ResizingTextEdit : public QTextEdit
 {
 public:
-    ResizingTextEdit()
-        : keyPressed()
-    {
-        auto sizePolicy = this->sizePolicy();
-        sizePolicy.setHeightForWidth(true);
-        sizePolicy.setVerticalPolicy(QSizePolicy::Preferred);
-        this->setSizePolicy(sizePolicy);
+    ResizingTextEdit();
 
-        QObject::connect(this, &QTextEdit::textChanged, this, &QWidget::updateGeometry);
-    }
+    QSize sizeHint() const override;
 
-    QSize sizeHint() const override
-    {
-        return QSize(this->width(), this->heightForWidth(this->width()));
-    }
-
-    bool hasHeightForWidth() const override
-    {
-        return true;
-    }
+    bool hasHeightForWidth() const override;
 
     boost::signals2::signal<void(QKeyEvent *)> keyPressed;
 
 protected:
-    int heightForWidth(int) const override
-    {
-        auto margins = this->contentsMargins();
+    int heightForWidth(int) const override;
+    void keyPressEvent(QKeyEvent *event) override;
 
-        return margins.top() + document()->size().height() + margins.bottom() + 5;
-    }
-
-    void keyPressEvent(QKeyEvent *event) override
-    {
-        event->ignore();
-
-        keyPressed(event);
-
-        if (!event->isAccepted()) {
-            QTextEdit::keyPressEvent(event);
-        }
-    }
+private:
+    QCompleter completer;
 };
